requirements.txt
flask
gunicorn
line-bot-sdk
psycopg2
flask-sqlalchemy
flask-cors
python-dotenv
